+++ packages/typespec-rust/.scripts/tspcompile.js
@@ -120,6 +120,9 @@ function should_generate(name) {
   return true
 }
 
+const appconfiguration = pkgRoot + 'test/tsp/AppConfiguration';
+generate('appconfiguration', appconfiguration, 'test/sdk/appconfiguration');
+
 const keyvault_secrets = pkgRoot + 'test/tsp/Security.KeyVault.Secrests';
 generate('keyvault_secrets', keyvault_secrets, 'test/sdk/keyvault_secrets');